Repository: introduce IsBare property
Introduce Repository.IsBare to determine if a repository is a bare
repository (without a working directory).

Additionally, introduce helper methods to ensure that we respect
disposed objects and do not attempt to invoke native functions on
freed or null native objects, and that we keep-alive IDisposeables
that contain references to native objects so that they are not GC'd
in mid-native execution.