Enable targeting .NET Standard 2.0
Enable targeting .NET Standard 2.0 as well as .NET Standard 2.1 to
allow consumers on .NET Framework and downlevel .NET Core 2.0 runtimes
to consume Dogged.

When the Target Framework is netstandard2.0 we pull in the System.Memory
package which includes the `Span<T>` and `Memory<T>` types that are used
in the library but missing from the standard < v2.1.