Lazy: specify factory in getter instead of ctor
Specifying the factory in the lazy's constructor means that one has an
object constructor with all the business logic. Moving it into a getter
means that the business logic for how to resolve a value from a native
resides _in the getter_. This shrinks the object constructor
considerably and makes the lazy data much easier to locate and reason
about.